chore: refactor code to module
