+++ src/chromedb/ccl_blink_value_deserializer.py
@@ -25,7 +25,7 @@ import enum
 import typing
 from dataclasses import dataclass
 
-import ccl_v8_value_deserializer
+from chromedb import ccl_v8_value_deserializer
 
 # See: https://chromium.googlesource.com/chromium/src/third_party/+/master/blink/renderer/bindings/core/v8/serialization
 #      https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/bindings/modules/v8/serialization/v8_script_value_serializer_for_modules.cc
