+++ src/chromedb/ccl_leveldb.py
@@ -31,7 +31,7 @@ import enum
 from collections import namedtuple
 from types import MappingProxyType
 
-import ccl_simplesnappy
+from chromedb import ccl_simplesnappy
 
 __version__ = "0.4"
 __description__ = "A module for reading LevelDB databases"
