+++ src/chromedb/ccl_simplesnappy.py
