+++ tools/benchmark.py
@@ -1,6 +1,6 @@
 import sys
 import pathlib
-import ccl_chromium_indexeddb
+from ..src.chromedb import ccl_chromium_indexeddb
 import time
