+++ tools/dump_leveldb.py
@@ -1,6 +1,6 @@
 import sys
 import csv
-import ccl_leveldb
+from ..src.chromedb import ccl_leveldb
 import pathlib
 
 ENCODING = "iso-8859-1"
